SpiceParser: iterate kwargs in pairs split on =
Instead of splitting on space and finding spaces, split on = so that we're guaranteed to get left and right halves of each expression, even when the expression has spaces in it.

The real problem here is that kwarg parsing doesn't match up with SPICE syntax. This is a step closer, but it's not likely perfect either.